Moved death cleanup system to the proper location
