+++ server/src/systems/handlers/game/mod.rs
@@ -1,7 +1,5 @@
 //! Event handlers for in-game events
 
-mod onkilledcleanup;
-
 mod register;
 
 pub mod on_chat_throttled;
@@ -18,5 +16,3 @@ pub mod on_spectate_event;
 pub mod timer;
 
 pub use self::register::register;
-
-pub use self::onkilledcleanup::PlayerKilledCleanup;
