+++ server/src/systems/handlers/game/register.rs
@@ -16,6 +16,7 @@ pub fn register<'a, 'b>(builder: Builder<'a, 'b>) -> Builder<'a, 'b> {
 		.with_handler::<on_player_killed::DisplayMessage>()
 		.with_handler::<on_player_killed::UpdateScore>()
 		.with_handler::<on_player_killed::CreateDespawnEvent>()
+		.with_handler::<on_player_killed::DespawnMissile>()
 		// On player joined
 		.with::<on_join::InitConnection>()
 		.with::<on_join::InitKillCounters>()
@@ -50,8 +51,6 @@ pub fn register<'a, 'b>(builder: Builder<'a, 'b>) -> Builder<'a, 'b> {
 		.with_handler::<on_player_respawn::SendPlayerRespawn>()
 		.with_handler::<on_player_respawn::CreateDespawnEvent>()
 		.with_handler::<on_player_respawn::GiveShield>()
-		// Misc
-		.with::<PlayerKilledCleanup>()
 		// Chat throttling
 		.with_registrar(on_chat_throttled::register)
 		// Timer events
